+++ solr/core/src/java/org/apache/solr/handler/admin/api/SnapshotBackupAPI.java
@@ -16,37 +16,33 @@
  */
 package org.apache.solr.handler.admin.api;
 
-import static org.apache.solr.client.solrj.impl.BinaryResponseParser.BINARY_CONTENT_TYPE_V2;
 import static org.apache.solr.handler.ReplicationHandler.ERR_STATUS;
 import static org.apache.solr.security.PermissionNameProvider.Name.CORE_EDIT_PERM;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import jakarta.inject.Inject;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 import java.util.function.Consumer;
 import org.apache.solr.api.JerseyResource;
-import org.apache.solr.client.api.model.SolrJerseyResponse;
+import org.apache.solr.client.api.endpoint.ReplicationBackupApis;
+import org.apache.solr.client.api.model.ReplicationBackupRequestBody;
+import org.apache.solr.client.api.model.ReplicationBackupResponse;
 import org.apache.solr.common.SolrException;
-import org.apache.solr.common.annotation.JsonProperty;
 import org.apache.solr.common.util.NamedList;
 import org.apache.solr.core.SolrCore;
 import org.apache.solr.handler.ReplicationHandler;
 import org.apache.solr.handler.ReplicationHandler.ReplicationHandlerConfig;
-import org.apache.solr.jersey.JacksonReflectMapWriter;
 import org.apache.solr.jersey.PermissionName;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-/** V2 endpoint for Backup API used for User-Managed clusters and Single-Node Installation. */
-@Path("/cores/{coreName}/replication/backups")
-public class SnapshotBackupAPI extends JerseyResource {
+/**
+ * v2 API implementation for replication-handler based backup creation.
+ *
+ * <p>This is the main backup functionality available to 'standalone' users.
+ */
+public class SnapshotBackupAPI extends JerseyResource implements ReplicationBackupApis {
 
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
   private final SolrCore solrCore;
@@ -62,22 +58,17 @@ public class SnapshotBackupAPI extends JerseyResource {
    * This API (POST /api/cores/coreName/replication/backups {...}) is analogous to the v1
    * /solr/coreName/replication?command=backup
    */
-  @POST
-  @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML, BINARY_CONTENT_TYPE_V2})
-  @Operation(summary = "Backup command using ReplicationHandler")
+  @Override
   @PermissionName(CORE_EDIT_PERM)
-  public BackupReplicationResponse createBackup(
-      @RequestBody BackupReplicationRequestBody backupReplicationPayload) throws Exception {
+  public ReplicationBackupResponse createBackup(
+      @RequestBody ReplicationBackupRequestBody backupReplicationPayload) {
     ensureRequiredRequestBodyProvided(backupReplicationPayload);
-    ReplicationHandler replicationHandler =
-        (ReplicationHandler) solrCore.getRequestHandler(ReplicationHandler.PATH);
-    return doBackup(replicationHandler, backupReplicationPayload);
+    return doBackup(backupReplicationPayload);
   }
 
-  private BackupReplicationResponse doBackup(
-      ReplicationHandler replicationHandler,
-      BackupReplicationRequestBody backupReplicationPayload) {
-    BackupReplicationResponse response = instantiateJerseyResponse(BackupReplicationResponse.class);
+  private ReplicationBackupResponse doBackup(
+      ReplicationBackupRequestBody backupReplicationPayload) {
+    ReplicationBackupResponse response = instantiateJerseyResponse(ReplicationBackupResponse.class);
     int numberToKeep = backupReplicationPayload.numberToKeep;
     int numberBackupsToKeep = replicationHandlerConfig.getNumberBackupsToKeep();
     String location = backupReplicationPayload.location;
@@ -128,63 +119,8 @@ public class SnapshotBackupAPI extends JerseyResource {
         resultConsumer);
   }
 
-  /* POJO for v2 endpoints request body. */
-  public static class BackupReplicationRequestBody implements JacksonReflectMapWriter {
-
-    public BackupReplicationRequestBody() {}
-
-    public BackupReplicationRequestBody(
-        String location, String name, int numberToKeep, String repository, String commitName) {
-      this.location = location;
-      this.name = name;
-      this.numberToKeep = numberToKeep;
-      this.repository = repository;
-      this.commitName = commitName;
-    }
-
-    @Schema(description = "The path where the backup will be created")
-    @JsonProperty
-    public String location;
-
-    @Schema(description = "The backup will be created in a directory called snapshot.<name>")
-    @JsonProperty
-    public String name;
-
-    @Schema(description = "The number of backups to keep.")
-    @JsonProperty
-    public int numberToKeep;
-
-    @Schema(description = "The name of the repository to be used for e backup.")
-    @JsonProperty
-    public String repository;
-
-    @Schema(
-        description =
-            "The name of the commit which was used while taking a snapshot using the CREATESNAPSHOT command.")
-    @JsonProperty
-    public String commitName;
-  }
-
-  /** Response for {@link SnapshotBackupAPI#createBackup(BackupReplicationRequestBody)}. */
-  public static class BackupReplicationResponse extends SolrJerseyResponse {
-
-    @JsonProperty("result")
-    public NamedList<?> result;
-
-    @JsonProperty("status")
-    public String status;
-
-    @JsonProperty("message")
-    public String message;
-
-    @JsonProperty("exception")
-    public Exception exception;
-
-    public BackupReplicationResponse() {}
-  }
-
   private static void reportErrorOnResponse(
-      BackupReplicationResponse response, String message, Exception e) {
+      ReplicationBackupResponse response, String message, Exception e) {
     response.status = ERR_STATUS;
     response.message = message;
     if (e != null) {
